Clean: Made it so archers are using the base enemy class vars
It wasnt using them, now they are for speed

+++ 2D_TurnBased_Template/Assets/Code/Base Classes/BaseEnemy.cs
@@ -1,11 +1,10 @@
 using UnityEngine;
-using UnityEngine.Rendering;
 
 public class BaseEnemy : MonoBehaviour
 {
     [Header("Enemy Stats")]
     public float EnemyHealth;
-    public int EnemySpeed;
+    public float EnemySpeed;
     public int EnemyDamage;
 
     [Header("Enemy Souls Value")]
